Move summarize by time column drill
